Fixed issues highlighted by make check
