+++ domain/src/main/kotlin/es/pedrazamiguez/splittrip/domain/usecase/balance/impl/GetMemberBalancesFlowUseCaseImpl.kt
@@ -8,8 +8,8 @@ import es.pedrazamiguez.splittrip.domain.model.SettlementRecord
 import es.pedrazamiguez.splittrip.domain.model.SettlementStatus
 import es.pedrazamiguez.splittrip.domain.service.AddOnCalculationService
 import es.pedrazamiguez.splittrip.domain.usecase.balance.GetMemberBalancesFlowUseCase
-import es.pedrazamiguez.splittrip.domain.usecase.balance.MemberBalanceCalculationInputs
 import es.pedrazamiguez.splittrip.domain.usecase.balance.support.ExpenseResult
+import es.pedrazamiguez.splittrip.domain.usecase.balance.support.MemberBalanceCalculationInputs
 import es.pedrazamiguez.splittrip.domain.usecase.balance.support.RemainingResult
 import es.pedrazamiguez.splittrip.domain.usecase.balance.support.WithdrawalResult
 import es.pedrazamiguez.splittrip.domain.usecase.balance.support.attributeExpensesByPaymentMethod
