+++ features/settlements/src/main/kotlin/es/pedrazamiguez/splittrip/features/settlement/presentation/component/PocketCashTile.kt
@@ -1,0 +1,54 @@
+package es.pedrazamiguez.splittrip.features.settlement.presentation.component
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.RowScope
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.unit.dp
+import es.pedrazamiguez.splittrip.core.designsystem.foundation.spacing
+import es.pedrazamiguez.splittrip.core.designsystem.presentation.component.layout.FlatCard
+import es.pedrazamiguez.splittrip.core.designsystem.presentation.component.text.AmountText
+import es.pedrazamiguez.splittrip.core.designsystem.presentation.component.text.CaptionText
+
+@Composable
+internal fun RowScope.PocketCashTile(
+    icon: ImageVector,
+    label: String,
+    formattedAmount: String,
+    amountColor: Color = MaterialTheme.colorScheme.onSurface
+) {
+    FlatCard(
+        modifier = Modifier.weight(1f)
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(MaterialTheme.spacing.Medium),
+            verticalArrangement = Arrangement.spacedBy(MaterialTheme.spacing.ExtraSmall)
+        ) {
+            Row(
+                horizontalArrangement = Arrangement.spacedBy(MaterialTheme.spacing.ExtraSmall),
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Icon(
+                    imageVector = icon,
+                    contentDescription = null,
+                    modifier = Modifier.size(14.dp),
+                    tint = MaterialTheme.colorScheme.onSurfaceVariant
+                )
+                CaptionText(text = label)
+            }
+            AmountText(text = formattedAmount, color = amountColor)
+        }
+    }
+}
